backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
ultralytics>=8.0.200
aiosqlite>=0.19.0
Pillow>=10.0.0
numpy>=1.26.0
supabase>=2.0.0
python-dotenv>=1.0.0
# zeroconf>=0.131.0  # Only needed for local mDNS, not in cloud

